## [0.7.89] - 2026-02-27 ### Fixed - **Build System**:   - Excluded `res.rc` from test executable build in headless CI mode, fixing a Ninja/RC compiler bug on Windows.   - Resolved `cmath` missing header errors by incorporating Visual Studio developer environment initialization `Launch-VsDevShell.ps1` within the local `clang-tidy` command flow. - **Static Analysis Fixes**:   - Addressed `bugprone-integer-division` issues in GUI layer latency calculations.   - Handled `bugprone-branch-clone` by merging duplicating branches in configuration parsing and yaw kick thresholding.   - Eliminated `performance-unnecessary-copy-initialization` on config versions.   - Handled `bugprone-empty-catch` logging errors rather than swallowing silently. - **CI Synchronization**:   - Unified `clang-tidy` exception flags in `windows-build-and-test.yml` to exactly match `build_commands.txt`.

+++ src/GuiLayer_Common.cpp
@@ -620,7 +620,7 @@ void GuiLayer::DrawTuningWindow(FFBEngine& engine) {
             if (ImGui::IsItemDeactivatedAfterEdit()) Config::Save(engine);
 
             ImGui::SameLine();
-            float latency_ms = (float)(engine.m_slope_sg_window / 2) * 2.5f;
+            float latency_ms = (engine.m_slope_sg_window / 2.0f) * 2.5f;
             ImVec4 color = (latency_ms < 25.0f) ? ImVec4(0,1,0,1) : ImVec4(1,0.5f,0,1);
             ImGui::TextColored(color, "~%.0f ms latency", latency_ms);
             ImGui::NextColumn(); ImGui::NextColumn();
